Move CriticalPoint class to eigenvalues module
